ConvertCwarToGoldCorpus/prepare-cwar.sh: Major rewrite, make it more robust and include files in sorted order

+++ bin/prepare-cwar.sh
@@ -5,10 +5,25 @@ if [ -z $FIELDSPRING_DIR ]; then
     exit
 fi
 
-origcwarxmldir=${1%/}
-pathtokml=$2
-pathtogaz=$3
-cwarxmloutdir=${4%/}
+origcwarxmldir=
+pathtokml=
+pathtotgn=
+pathtogaz=
+cwarxmloutdir=
+keeptmp=no
+
+# Handle command-line args for us.
+while true ; do
+  case "$1" in
+    -c | --corpus ) origcwarxmldir="${2%/}"; shift 2 ;;
+    -k | --kml ) pathtokml="$2"; shift 2 ;;
+    -t | --tgn ) pathtotgn="$2"; shift 2 ;;
+    -g | --gaz ) pathtogaz="$2"; shift 2 ;;
+    -o | --outdir ) cwarxmloutdir="${2%/}"; shift 2 ;;
+    --keep-tmp ) keeptmp=yes; shift ;;
+    * ) break ;;
+  esac
+done
 
 echo "Converting original Cwar corpus to plain format..."
 cwarxml2txttgn.sh $origcwarxmldir cwarplaintgn
@@ -24,14 +39,35 @@ if [ ! -e $cwarxmloutdir/test ]; then
     mkdir $cwarxmloutdir/test
 fi
 
+pathtotgncvt=
+coordargs=
+if [ -n "$pathtotgn" ]; then
+  echo "Converting TGN coordinates to a simple text file..."
+  pathtotgncvt=tgn-coords.$$.txt
+  if echo "$pathtotgn" | grep '\.bz2$' > /dev/null ; then
+    bzcat "$pathtotgn" | convert-TGNOut_Coordinates.nt.sh > $pathtotgncvt
+  else
+    convert-TGNOut_Coordinates.nt.sh < "$pathtotgn" > $pathtotgncvt
+  fi
+  coordargs="$coordargs -t $pathtotgncvt"
+fi
+if [ -n "$pathtokml" ]; then
+  coordargs="$coordargs -k $pathtokml"
+fi
+
 echo "Converting dev corpus to Fieldspring format..."
-fieldspring --memory 8g run opennlp.fieldspring.tr.app.ConvertCwarToGoldCorpus cwarplaintgndev $pathtokml $pathtogaz > $cwarxmloutdir/dev/cwar-dev.xml
+fieldspring --memory 15g run opennlp.fieldspring.tr.app.ConvertCwarToGoldCorpus -c cwarplaintgndev $coordargs -g $pathtogaz > $cwarxmloutdir/dev/cwar-dev.xml
 echo "Converting test corpus to Fieldspring format..."
-fieldspring --memory 8g run opennlp.fieldspring.tr.app.ConvertCwarToGoldCorpus cwarplaintgntest $pathtokml $pathtogaz > $cwarxmloutdir/test/cwar-test.xml
+fieldspring --memory 15g run opennlp.fieldspring.tr.app.ConvertCwarToGoldCorpus -c cwarplaintgntest $coordargs -g $pathtogaz > $cwarxmloutdir/test/cwar-test.xml
 
-echo "Deleting temporary files..."
-rm -rf cwarplaintgn
-rm -rf cwarplaintgndev
-rm -rf cwarplaintgntest
+if [ "$keeptmp" != "yes" ]; then
+  echo "Deleting temporary files..."
+  rm -rf cwarplaintgn
+  rm -rf cwarplaintgndev
+  rm -rf cwarplaintgntest
+  if [ -n "$pathtotgncvt" ]; then
+    rm -rf $pathtotgncvt
+  fi
+fi
+
 echo "Done."
-
